Move `FixedRate` into event_loop module
This is where these types are used, they can be defined in there.
