+++ swap/src/asb.rs
@@ -1,7 +1,5 @@
 pub mod command;
 pub mod config;
-mod fixed_rate;
 mod rate;
 
-pub use self::fixed_rate::FixedRate;
-pub use self::rate::Rate;
+pub use rate::Rate;
