+++ swap/tests/harness/mod.rs
@@ -15,11 +15,11 @@ use std::fmt;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use std::time::Duration;
-use swap::asb::FixedRate;
 use swap::bitcoin::{CancelTimelock, PunishTimelock};
 use swap::database::Database;
 use swap::env::{Config, GetConfig};
 use swap::network::swarm;
+use swap::protocol::alice::event_loop::FixedRate;
 use swap::protocol::alice::{AliceState, Swap};
 use swap::protocol::bob::BobState;
 use swap::protocol::{alice, bob};
